# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Subcults API
  description: |
    Privacy-first platform for mapping underground music communities.
    Combines real-time location awareness, live audio streaming, and trust-based discovery.

    ## Authentication
    JWT-based authentication with two token types:
    - **Access tokens**: 15-minute expiry, includes `did` (Decentralized Identifier) claim
    - **Refresh tokens**: 7-day expiry, cookie-only

    Pass the access token via `Authorization: Bearer <token>` header or the `access_token` cookie.

    ## Privacy
    All public location data uses geohash-based jitter to prevent precise location tracking.
    Precise coordinates are only available when `allow_precise` consent is granted.

    ## Rate Limiting
    Per-endpoint rate limits are enforced. When exceeded, the API returns `429 Too Many Requests`
    with a `Retry-After` header.
  version: 0.1.0
  contact:
    name: Subcults
  license:
    name: MIT
servers:
  - url: /
    description: Current deployment
...
paths:
...
  /alliances:
    post:
      operationId: createAlliance
      tags: [Alliances]
      summary: Create a trust alliance between two scenes
      security:
        - bearerAuth: []
      x-rate-limit: 10/hour
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAllianceRequest'
      responses:
        '201':
          description: Alliance created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alliance'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Self-alliance or duplicate
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          $ref: '#/components/responses/RateLimited'
  /alliances/{id}:
    parameters:
      - $ref: '#/components/parameters/ResourceID'
    get:
      operationId: getAlliance
      tags: [Alliances]
      summary: Get an alliance by ID
      security:
        - bearerAuth: []
        - {}
      responses:
        '200':
          description: Alliance found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alliance'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateAlliance
      tags: [Alliances]
      summary: Update an alliance
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAllianceRequest'
      responses:
        '200':
          description: Alliance updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alliance'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteAlliance
      tags: [Alliances]
      summary: Delete an alliance
      security:
        - bearerAuth: []
      responses:
        '204':
          description: Alliance deleted
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

  # ── Trust ───────────────────────────────────────────────────────────
...
  /api/csp-report:
    post:
      operationId: cspReport
      tags: [Health]
      summary: CSP violation report endpoint
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '204':
          description: Report received
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'JWT access token. Include as Authorization: Bearer <token> header or access_token cookie.'
  parameters:
    ResourceID:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds until the rate limit resets
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              enum:
                - validation_error
                - auth_failed
                - not_found
                - rate_limited
                - internal_error
                - forbidden
                - conflict
                - bad_request
                - invalid_palette
                - scene_deleted
                - scene_not_found
                - invalid_scene_name
                - duplicate_scene_name
                - invalid_time_range
                - missing_target
                - unsupported_type
                - invalid_weight
                - alliance_deleted
                - self_alliance
                - already_onboarded
                - not_onboarded
            message:
              type: string
              description: Human-readable error message

    # ── Geographic ──────────────────────────────────────────────────
...
    Alliance:
      type: object
      required: [from_scene_id, to_scene_id, weight]
      properties:
        from_scene_id:
          type: string
        to_scene_id:
          type: string
        weight:
          type: number
          format: double
          minimum: 0
          maximum: 1
        reason:
          type: string
          maxLength: 256
    CreateAllianceRequest:
      type: object
      required: [from_scene_id, to_scene_id, weight]
      properties:
        from_scene_id:
          type: string
        to_scene_id:
          type: string
        weight:
          type: number
          format: double
          minimum: 0
          maximum: 1
        reason:
          type: string
          maxLength: 256
    UpdateAllianceRequest:
      type: object
      properties:
        weight:
          type: number
          format: double
          minimum: 0
          maximum: 1
        reason:
          type: string
          maxLength: 256

    # ── Trust ───────────────────────────────────────────────────────
...